kevinobriendotca/ARO-17938-Check-Gateway-VMSS-Instance-State-Before-Stopping - Updated logic in isVMInstanceHealthy to OR the conditions to catch nils along the way to the actual health status code.

+++ pkg/deploy/upgrade_gateway.go
@@ -80,12 +80,15 @@ func (d *deployer) gatewayRemoveOldScaleset(ctx context.Context, vmssName string
 	d.log.Printf("stopping scaleset %s", vmssName)
 	errors := make(chan error, len(scalesetVMs))
 	for _, vm := range scalesetVMs {
-		go func(id string) {
-			errors <- d.vmssvms.RunCommandAndWait(ctx, d.config.GatewayResourceGroupName, vmssName, id, mgmtcompute.RunCommandInput{
-				CommandID: to.StringPtr("RunShellScript"),
-				Script:    &[]string{"systemctl stop aro-gateway"},
-			})
-		}(*vm.InstanceID) // https://golang.org/doc/faq#closures_and_goroutines
+		if d.isVMInstanceHealthy(ctx, d.config.GatewayResourceGroupName, vmssName, *vm.InstanceID) {
+			d.log.Printf("stopping gateway service on %s", *vm.Name)
+			go func(id string) {
+				errors <- d.vmssvms.RunCommandAndWait(ctx, d.config.GatewayResourceGroupName, vmssName, id, mgmtcompute.RunCommandInput{
+					CommandID: to.StringPtr("RunShellScript"),
+					Script:    &[]string{"systemctl stop aro-gateway"},
+				})
+			}(*vm.InstanceID) // https://golang.org/doc/faq#closures_and_goroutines
+		}
 	}
 
 	d.log.Print("waiting for instances to stop")
